pyproject.toml
[project]
name = "bracket-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "ecdsa>=0.19.0",
    "fastapi>=0.115.8",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "uvicorn>=0.33.0",
]
